Build 64bit binaries for recovery and userfastboot on 32p builds
     32p builds set the primary ABI to 64bit and secondary
ABI to 32bit but also set a flag to prefer 32bit binaries
be built. This works well for the system image but causes
problem for the recovery and userfastboot images. For these
images, we need to use 64bit for all dynamic binaries and
libraries. We use the LOCAL_MULTILIB feature to force the
binaries being used in these images to be 64bit in this case.
The 32o image will continue to have 32bit binaries.

Change-Id: If927dccf3f92d5fe89209db968b0fe935185db5b
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Reviewed-by: Shi, Qiming <qiming.shi@intel.com>