Update patch set 1
Patch Set 1:

It doesn't copy any headers. It adds a "-I external/zlib" line to the export_includes in the libz_intermediates  directory this is picked up by any projects that depend on the libz. This means you only need to add the relevant LOCAL_SHARED_LIBRARIES / LOCAL_STATIC_LIBRARIES entry and not the LOCAL_C_INCLUDES line too. 

I assume it's intention is to make the whole thing more portable and allows you to move/rename the zlib directory without breaking dependent projects .... well not so in the case of zlib as a number of projects have a LOCAL_C_INCLUDES := external/zlib but that an be "fixed" should this changed be accepted

Patch-set: 1