Update patch set 1
Patch Set 1:

(1 comment)

OOI - What stats would be incorrect?

I assume an overflow doesn't break anything for the actual crc calcs...? Is the problem that crc is being used for something else that requires that it doesn't overflow? Some kind of size calc?

<Usual caveats here about being a Java guy that doesn't understand native code>:

I assume there are other changes coming to keep it all compiling. e.g. those places that are using long crc = crc32(); or int crc =, etc.

Is there a bug number for the Deflate not working?

Patch-set: 1