Update patch set 1
Patch Set 1:

if i build this, i'm told to rebuild the abi dumps. but if i do that, it creates a new "Q" directory in the prebuilts. which doesn't seem right :-)

is there some documentation about what i should do in a case like this?

Patch-set: 1
CC: Gerrit User 1014807 <1014807@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>