Update patch set 1
Patch Set 1:

I am sort of surprised by the files under prebuilts/abi-dumps/vndk, but I think both of them are correct (correspond to the NDK version and the VNDK version respectively).

However, this reminds me that the way the build system looks for the reference ABI dumps should be revised for this possibility.  I'll file another bug to fix that.

Patch-set: 1