Update patch set 1
Patch Set 1: Code-Review+1

@Tao: thanks for adding me to the patch, I really appreciate it.

If I understood correctly, it will allow to run a test that relies on imgdiff? If so, that is really handy, it is a good idea!

I'm not an android maintainer (only Chromium's zlib owner), but it looks good.

Patch-set: 1
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1