Update patch set 5
Patch Set 5:

> 
> +calin@ since i'm not sure whether we'd expect this to pass if we change the zip CRC. (the fact that the array is a different _size_ is a little worrying though...)

Indeed.

This seems oddly similar to the case of GMeet when we first landed the inflate_fast optimization that I mentioned on the issue (https://issuetracker.google.com/issues/115695768).

Quote:
"
...
d) Tales from the trenches: when we first landed the inflate_fast wide loads/stores optimization in 2017, there was a regression with an App (Google Meeting) as the user authentication was no longer working. For context, Chromium's zlib is used by Cronet (Chromium Network Library) that is used by quite a few Android apps.

The issue was with GMeet app Java code as it was relying on non-specified behavior of vanilla zlib (i.e. zero the whole input buffer) that has changed due to  the optimization (i.e. we may perform a long store and write some garbage at the end of the buffer, but we keep track of the correct size of decompressed content and also ensure we got enough room to perform the write).

The public API provides an argument that lets the user known what is the decompressed content size, *but* GMeet instead kept reading the buffer until it found a zero...
"

Question: what would be the build target/test name to reproduce the failure?

Patch-set: 5
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>