Update patch set 5
Patch Set 5:

I'm fine with the change. 

I suspect the test fails because the new optimizations change the actual bytes which result after ziping. In the tests we compare (on byte level) a prebuilt zip archive with the a zip-file produced by some system APIs (snapshot-profile).

I think that should be fixable at the test level (either update the prebuilds or test the archive content instead of the byte-level). However, I have no cycles to update the tests in the immediate future so feel free to upload a fix to them as well to satisfy tree hugger and keep everything green.

Patch-set: 5