Update patch set 1
Patch Set 1:

unfortunately it turns out that the libcore tests aren't in presubmit? looks like in addition to the NDK breakage (which is just because the NDK still needs to support non-Neon, and is easily worked around), the zlib optimizations break a bunch of libcore tests.

they sound pretty scary:

  org.apache.harmony.tests.java.util.zip.DeflaterTest#test_ConstructorIZ
    junit.framework.AssertionFailedError: final decompressed data contained more bytes than original - constructorIZ expected:<0> but was:<85>

...but when you actually look at the test, they're not actually looking at how many bytes the inflater claims to have returned. (and are surprised to see that the inflater has actually touched more bytes than expected in the output array.)

        byte outPutInf[] = new byte[500];
        Inflater infl = new Inflater(true);
        while (!infl.finished()) {
            if (infl.needsInput()) {
                infl.setInput(outPutBuf);
            }
            infl.inflate(outPutInf);
        }
        for (int i = 0; i < byteArray.length; i++) {
            assertEquals(byteArray[i], outPutInf[i]);
        }
        assertEquals("final decompressed data contained more bytes than original - constructorIZ",
                0, outPutInf[byteArray.length]);

(most failures appear to be copy & pastes of this, but there's also one where a deflate using the huffman strategy outputs 99 bytes instead of 100.)

Patch-set: 1
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>