Update patch set 3
Patch Set 3:

> > Since I'm back from vacations, I should have time to help out on this one.
> > :-)
> 
> yeah, turning off the inflate optimization (as in the latest patchset) fixes the tests. so i think we can at least get the other optimizations in. but first i need to verify that i haven't broken the NDK again...

It is pretty nice that the hunch was correct, and thanks for validating the theory i.e. INFLATE_CHUNK_SIMD_NEON would break the tests.

Maybe we could target for now to have all the optimizations enabled (with the exception of INFLATE_CHUNK) and in the near future work to fix the failing tests since they are relying on an unspecified behavior?

The optimizations currently enabled should improve decompression performance in average 70-80% and compression around 30-40% on ARM.

There is one extra consideration about INFLATE_CHUNK: that just like the tests, user apps may be relying on this unspecified behavior (i.e. not actually looking at how many bytes the inflater claims to have returned, but expect some invariant on the output buffer tail) and could potentially break if INFLATE_CHUNK is enabled.

In Chromium we have a deprecation process for letting users known when a user-facing feature is going to change, so the users can have 1 or 2 releases to adjust their websites for the upcoming changes.

I wonder if there is something similar for Android releases?

Patch-set: 3
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>