Update patch set 3
Patch Set 3:

here's the failure in CtsDexMetadataHostTestCases with what this CL enables in zlib:

expected: [112, 114, 111, 0, 48, 49, 48, 0, 1, -115, 0, 0, 0, 41, 0, 0, 0, 120, 1, -29, 96, 48, 102, 0, -127, -49, -121, -66, -88, 89, 1, -23, -92, -60, -30, 84, -67, -60, -126, 108, 94, 6, 70, 58, -64, 3, -1, -63, -128, 17, 66, -79, 3, 0, -99, 17, 19, 69]
actual:   [112, 114, 111, 0, 48, 49, 48, 0, 1, -115, 0, 0, 0, 42, 0, 0, 0, 120, 1, -29, 96, 48, 102, 0, -127, -49, -121, -66, -88, 89, 49, 48, 48, 36, 37, 22, -89, -22, 37, 22, 100, -13, 50, 48, -46, 1, 30, -8, 15, 6, -116, 16, -118, 29, 0, -99, 17, 19, 69]

i'm actually seeing this on arm64 too, so i don't know why we didn't get a bug filed for this last time we were merged.

anyway, with trusty `man ascii` as my guide, that looks like the header is the expected "pro\0""101\0", but then the zipped sequence of profile line headers and profile line data differs. i'll try to change the test to unzip the data after the headers and compare that instead, which seems like the most robust fix. (i'm assuming that things are okay if we do that, but i haven't actually confirmed it yet.)

Patch-set: 3
CC: Gerrit User 1022077 <1022077@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0