Update patch set 3
Patch Set 3:

I got a few more questions related to updates and security fixes.

The questions are:
a) What is the policy for updates? For Chromium we have both a shorter release cycle (i.e. next version will have most recent changes) and a special process to backport security fixes on older releases. 

b) Concerning the NDK: can a user link with a static version of zlib and in a scenario where there is a security fix, still be stuck using an outdated zlib version?

Patch-set: 3
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>