Update patch set 3
Patch Set 3: Presubmit-Ready+1

(sorry for the delay. between branching for the R DP1 release and the start of performance review season...)

> Patch Set 3:
> 
> I got a few more questions related to updates and security fixes.
> 
> The questions are:
> a) What is the policy for updates? For Chromium we have both a shorter release cycle (i.e. next version will have most recent changes) and a special process to backport security fixes on older releases. 

the usual Android policy. we'll track upstream to within a couple of weeks, and there will be one big 
"dessert" release year. security fixes are judged by severity. iirc we've never had a zlib security issue bad enough to warrant a backport, but theoretically one could be backported to the last three releases. (there's a lower bar for backports to the _current_ shipping release, but that wasn't true last time there was a zlib security issue.)

if/when we actually make the runtime module mandatory and update libc/libm/libdl via the Play Store, i think we'll have to include libz too given that it's also a core NDK API. but it's potentially pretty disruptive, as we've seen here.

> b) Concerning the NDK: can a user link with a static version of zlib and in a scenario where there is a security fix, still be stuck using an outdated zlib version?

yes, an NDK user can statically link zlib. on the bright side, because there's a libz.a in the NDK they can conversely ship a *new* zlib even to an ancient device that isn't getting updates. (this is one reason why the security folks wish that libz.so wasn't part of the NDK and everyone was forced to use libz.a!) in practice, app developers are so sensitive about apk size that i doubt many are actually statically linking zlib in the wild.

anyway, let's kick presubmit again and see if we can get this merged... (i've got a few other things on the go at the moment, or i'd also see whether we can re-enable the stuff that's currently disabled. that can come later.)

Patch-set: 3
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1