zlib.h: annotate late arrivals with their API level.
I'm sure I'll live to regret this, but it does seem like our least worst
choice if we want to accurately represent the reality that different API
levels have different versions of zlib available.

Interestingly, today's archaeological dig turned up the fact that we
were wrong about when the 1.2.7.1 API appeared. It turns out that --
surprisingly -- 1.2.7 != 1.2.7.1 in terms of exposed API.

Note to self: move libz into the same apex as libc so we can get out of
this mess eventually, even if it takes another 10 years.

Bug: https://issuetracker.google.com/155351357
Bug: https://github.com/android/ndk/issues/1116
Test: builds
Change-Id: I77df497d89ce9e242458605fe5b08a3780c9d0f6