Update patch set 2
Patch Set 2:

Elliott

Help me understand the problem: you are trying to hide/show some functions depending on the Android API version?

Ideally we should *not* have to fork zlib.h only for AOSP, it would be preferable to have this on our upstream (i.e. Chromium) and I would be more than happy to help out here.

If that is not technically possible, I guess you could still have a copy of zlib.h on the Android repo.

From a quick look on the functions listed on this patch, I can say from the top of my head:
a) crc32_z and adler32_z *should* be alright, since they were introduced recently (IIRC zlib 1.2.9 onwards) and client code is *not* supposed to directly call them, instead use plain crc32() and adler32(). 

I recall that crc32() is used on network code and adler32() was being used in V8 as a generic checksum for V8 snapshots.

b) The *dictionary functions should be alright, as I think they are not actively used in Chromium codebase (but we got to double check).

c) inflateValidate() maybe be a problem as that unless I'm mistaken is used in libpng.

d) gzread/gzwrite IIRC was used in libxml. I dunno if that code is built or not, but there is only one way to make sure: we could upload this patch to Chromium infrastructure and kick the bots and see if they will turn red.

Patch-set: 2