Update patch set 2
Patch Set 2:

An update: I had a look on the testing coverage of Chromium's zlib and decided to improve it.

Code coverage for decompression now is around 96%:
https://bugs.chromium.org/p/chromium/issues/detail?id=1032721#c21

Any changes we land on Chromium's zlib are gated on those tests passing, which should improve the safety of our optimizations and development workflow.

I'm planning next to dive into the JNI glue code and see what could be the reason for the odd behavior exhibited by the failing Java test.

Patch-set: 2
Reviewer: Gerrit User 1512829 <1512829@85c56323-6fa9-3386-8a01-6480fb634889>